requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.3
djangorestframework==3.15.2
fpdf==1.7.2
idna==3.7
joblib==1.4.2
nltk==3.9.1
regex==2024.11.6
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
Uni-Curses==3.1.2
urllib3==2.2.2
x256==0.0.3
